+++ cuda_rasterizer/forward.h
@@ -16,6 +16,7 @@
 #include "cuda_runtime.h"
 #include "device_launch_parameters.h"
 #define GLM_FORCE_CUDA
+#include <cuda_fp16.h>
 #include <glm/glm.hpp>
 
 namespace FORWARD
@@ -56,10 +57,10 @@ namespace FORWARD
 		const float2* points_xy_image,
 		const float* features,
 		const float4* conic_opacity,
-		float* final_T,
+		half* final_T,
 		uint32_t* n_contrib,
 		const float* bg_color,
-		float* out_color);
+		half* out_color);
 }
